Move numa related fucntionality to own file
